Find a Mother Vertex in a Graph
What is a Mother Vertex? 
A mother vertex in a graph G = (V, E) is a vertex v such that all other vertices in G can be reached by a path from v
Note: There can be more than one mother vertices in a graph. We need to output anyone of them.